More WIP

+++ sbor/src/v1/constants.rs
@@ -1,4 +1,4 @@
 // Constants
 
 pub const SBOR_V1_PREFIX_BYTE: u8 = 0x5b; // 5b looks a bit like "Sb"(or)
-pub const DEFAULT_MAX_ENCODING_DEPTH: u8 = 32;
+pub const DEFAULT_MAX_ENCODING_DEPTH: u32 = 32;
